% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSSG.toolkit.R
\name{cell_stat_graph}
\alias{cell_stat_graph}
\title{Cell Type Statistics Graph}
\usage{
cell_stat_graph(data, include_ns = TRUE)
}
\arguments{
\item{data}{A data frame containing cell type counts and significance categories.}

\item{include_ns}{Logical. Whether to include non-significant cell types in the graph (default: TRUE).}
}
\value{
A ggplot object representing the bar plot.
}
\description{
Creates a bar plot of cell type distributions and significance categories.
}
\examples{
cell_stat_graph(data, include_ns = TRUE)

}
