% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSSG.toolkit.R
\name{create_project}
\alias{create_project}
\title{Create a scRNA Project from Sparse Matrix Files}
\usage{
create_project(
  sparse_matrix_path,
  sparse_name,
  rows_name,
  cols_name,
  type = "count"
)
}
\arguments{
\item{sparse_matrix_path}{A character string specifying the directory containing the sparse matrix files.}

\item{sparse_name}{The base name of the sparse matrix file (without the file extension, e.g., '.mtx').}

\item{rows_name}{The base name of the file containing the gene names (without the file extension, e.g., '.tsv').}

\item{cols_name}{The base name of the file containing the cell names (without the file extension, e.g., '.tsv').}

\item{type}{A character string specifying the type of data to read. This should be either 'count' or 'norm', depending on the data format.}
}
\value{
A new `scRNAproject` object containing the loaded sparse matrix and associated cell/genes names.
  - `matrices$count` or `matrices$norm`: The sparse matrix loaded from the provided path.
  - `names$primary`: The cell names extracted from the provided column names file.
}
\description{
This function initializes a custom `scRNAproject` object using a sparse matrix format for single-cell RNA-seq data. The function loads the expression data from the provided matrix and row/column name files. It also assigns the appropriate names to the cells and genes, based on the provided file paths.
}
\details{
The function reads the sparse matrix and its associated row and column name files (typically in `.tsv` format). It creates a custom `scRNAproject` object and assigns the data to appropriate slots for easy access and analysis.
}
\examples{
sc_project <- create_project("/path/to/data", "matrix_data", "genes", "cells")

}
