% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSSG.toolkit.R
\name{create_project_from_seurat}
\alias{create_project_from_seurat}
\title{Create a scRNA Project from a Seurat Object}
\usage{
create_project_from_seurat(seurat_project)
}
\arguments{
\item{seurat_project}{A Seurat object containing single-cell RNA-seq data, including normalized gene expression data.}
}
\value{
A new `scRNAproject` object containing the matrices and names extracted from the Seurat object. The `scRNAproject` object contains:
  - `matrices$norm`: The normalized expression matrix from the Seurat object.
  - `names$primary`: The cell identities from the Seurat object.
}
\description{
This function converts a Seurat object into a custom `scRNAproject` object. It extracts the normalized expression data from the Seurat object, assigns it to the `norm` slot in the matrices of the custom project, and sets the active identities as the names of the cells. This facilitates further analysis using custom functions designed for `scRNAproject` objects.
}
\examples{
sc_project <- create_project_from_seurat(seurat_project)

}
