% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSSG.toolkit.R
\name{cssg_naming}
\alias{cssg_naming}
\title{CSSG Cluster Naming}
\usage{
cssg_naming(sparse_matrix, CSSG_df, species = "Homo sapiens")
}
\arguments{
\item{sparse_matrix}{A sparse matrix containing single-cell gene expression data.}

\item{CSSG_df}{A data frame containing CSSG information, including loss values and adjustment metrics.}

\item{species}{A character string specifying the species (default: 'Homo sapiens').}
}
\value{
A character vector of cluster names.
}
\description{
Assigns names to clusters in a sc_project (sparse_matrix) object based on a CSSG data frame obtained using CSSG_markers() function.
}
\examples{
cssg_naming(sparse_matrix, CSSG_df)

}
