% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSSG.toolkit.R
\name{get_avg_data}
\alias{get_avg_data}
\title{Compute Averaged Expression Data}
\usage{
get_avg_data(sc_project, type = "subtypes", data = "norm", chunk_size = 5000)
}
\arguments{
\item{sc_project}{An object containing single-cell RNA-seq data.}

\item{type}{A string specifying the grouping type. Accepted values are 'subtypes' or other categorical groupings.}

\item{data}{A string specifying the data slot to use (e.g., 'norm' for normalized expression values). Default is 'norm'.}

\item{chunk_size}{An integer defining the chunk size for data aggregation. Default is 5000.}
}
\value{
A matrix or data frame with aggregated expression values.
}
\description{
This function retrieves and aggregates expression data from an `sc_project` object based on the specified cell type category.
}
\examples{
avg_data <- get_avg_data(sc_project, type = 'subtypes', data = 'norm')

}
