% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSSG.toolkit.R
\name{get_data}
\alias{get_data}
\title{Retrieve Specific Expression Data from an sc_project Object}
\usage{
get_data(sc_project, type = "subtypes", data = "norm")
}
\arguments{
\item{sc_project}{An object containing single-cell RNA-seq data.}

\item{type}{A string specifying the grouping type. Accepted values are 'subtypes' or 'subclasses'.}

\item{data}{A string specifying the type of expression data to retrieve. Accepted values are 'norm' (normalized) or 'count' (raw counts).}
}
\value{
A sparse matrix of the specified expression data, with appropriate cell names assigned based on the chosen `type`.
}
\description{
This function retrieves expression data (either 'norm' or 'count') from a specified subtype or subclass grouping within a given `sc_project` object.
}
\examples{
norm_data <- get_data(sc_project, type = "subtypes", data = "norm")
count_data <- get_data(sc_project, type = "subclasses", data = "count")

}
