% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSSG.toolkit.R
\name{get_names_markers}
\alias{get_names_markers}
\title{Retrieve Marker Gene Names}
\usage{
get_names_markers(sc_project, type = "subtypes")
}
\arguments{
\item{sc_project}{An object containing scRNA-seq data, including marker names.}

\item{type}{A string specifying the grouping type. Accepted values are 'subtypes', 'subclasses', 'cluster', or 'primary'. Default is 'subtypes'.}
}
\value{
A character vector of unique marker gene names that are present in the normalized expression matrix.
}
\description{
This function extracts marker gene names from a given `sc_project` object based on the specified cell subclass and subtype name.
}
\examples{
markers <- get_names_markers(sc_project, type = "subtypes")
print(markers)

}
