% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSSG.toolkit.R
\name{heterogeneity_select_variance}
\alias{heterogeneity_select_variance}
\title{Select Marker Genes Based on Expression Variance and Heterogeneity}
\usage{
heterogeneity_select_variance(
  sc_project,
  heterogeneity_factor = 80,
  max_genes = 1000,
  min_occ = 5,
  min_exp = 0.1,
  rep_factor = 0.1,
  mito_content = FALSE
)
}
\arguments{
\item{sc_project}{A single-cell project object containing normalized gene expression data.}

\item{heterogeneity_factor}{A numeric value defining the maximum percentage of cells in a
cluster expressing a gene for it to be considered heterogeneous. Default is 80.}

\item{max_genes}{A numeric value defining the maximum number of genes to retain per cluster.
Default is 1000.}

\item{min_occ}{A numeric value specifying the minimum percentage occurrence of a gene in a cluster.
Default is 5.}

\item{min_exp}{A numeric value specifying the minimum expression level threshold for considering
a gene in the variance calculation. Default is 0.1.}

\item{rep_factor}{A numeric value defining the replication threshold for filtering genes appearing
in multiple clusters. Default is 0.5.}

\item{mito_content}{A logical value indicating whether to include mitochondrial genes in the selection.
Default is FALSE.}
}
\value{
An updated `sc_project` object with selected heterogeneity marker genes stored in
  `metadata$heterogeneity_markers`.
}
\description{
Identifies cluster-specific marker genes using expression variance, occurrence thresholds,
and heterogeneity factors.
}
\details{
The function performs the following steps:
  - Extracts and processes normalized expression data.
  - Calculates gene variance and occurrence using `genes_variance_calculate()`.
  - Removes mitochondrial genes if `mito_content = FALSE`.
  - Filters genes based on heterogeneity constraints (`min_occ`, `heterogeneity_factor`).
  - Removes genes appearing in too many clusters based on `rep_factor`.
  - Selects the top `max_genes` per cluster based on expression variance.
  - Stores the final gene list in `sc_project@metadata$heterogeneity_markers`.
}
\examples{
updated_project <- heterogeneity_select_variance(sc_project, heterogeneity_factor = 50,
  max_genes = 200, min_occ = 10, min_exp = 0.2, rep_factor = 0.4)

}
