% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSSG.toolkit.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize Single-Cell RNA-seq Data}
\usage{
normalize_data(sc_project, type = "counts", factor = 1e+06)
}
\arguments{
\item{sc_project}{An object containing single-cell RNA-seq data with pre-existing count matrices.}

\item{type}{A string indicating whether to normalize by 'counts' or 'genes'. Defaults to 'counts'.}

\item{factor}{A numeric value used to scale the normalized data. Default is 1,000,000 (for TPM-like normalization).}
}
\value{
A modified `sc_project` object with normalized data stored in the `@matrices$norm` slot.
}
\description{
This function normalizes the count matrix of a single-cell RNA-seq project by applying a log transformation to the data. The normalization is based on either 'counts' or 'genes' and scales the data by a specified factor (default is 1,000,000). 
The log transformation is calculated using the formula: log2((count / gene counts) * factor + 1).
}
\examples{
normalized_project <- normalize_data(sc_project, type = 'counts', factor = 1000000)
normalized_project <- normalize_data(sc_project, type = 'genes', factor = 10000)

}
