% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSSG.toolkit.R
\name{reduce_extended_gtf}
\alias{reduce_extended_gtf}
\title{Reduce parts of gene names added by GTF-tool}
\usage{
reduce_extended_gtf(names)
}
\arguments{
\item{names}{A vector of gene names}
}
\value{
A vector of reduced gene names
}
\description{
This function modifies a vector of gene names by removing parts
added by GTF-tool (https://github.com/jkubis96/GTF-tool),
making it suitable for bulk analysis without distinguishing location
or transcript version.
}
\examples{
reduced_names <- reduce_extended_gtf(names)

}
