% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSSG.toolkit.R
\name{scale_reduced_names}
\alias{scale_reduced_names}
\title{Reduce parts of gene names added by GTF-tool and sum expression values}
\usage{
scale_reduced_names(data, chunk_size = 5000)
}
\arguments{
\item{data}{A data.frame, matrix, or sparse matrix of expression data}
}
\value{
A data.frame with adjusted expression values
}
\description{
This function modifies row names (genes) by removing parts
added by the GTF-tool (https://github.com/jkubis96/GTF-tool),
and sums expression values for genes with the same name after reduction.
This makes the data suitable for bulk analysis without distinguishing
gene location or transcript version.
}
\examples{
reduced_data <- scale_reduced_names(data, chunk_size = 5000)

}
