% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSSG.toolkit.R
\name{subclass_naming}
\alias{subclass_naming}
\title{Assign Subclass and Class Names to Cells in Single-Cell RNA-Seq Data}
\usage{
subclass_naming(
  sc_project,
  class_markers = NULL,
  subclass_markers = NULL,
  species = "Homo sapiens",
  chunk_size = 5000
)
}
\arguments{
\item{sc_project}{An object of class `scRNAproject`, which contains the data and metadata required for subclass and class naming. It should include matrices with normalized gene expression data and metadata related to heterogeneity markers.}

\item{species}{A character string specifying the species (default: 'Homo sapiens'), which determines the structure of the names.}

\item{chunk_size}{The number of cells to process in each chunk during aggregation (default is 5000).}

\item{markers_class}{A list of class markers (default: NULL). If NULL the cluster will obtain native names from 'sc_project@names$primary' or 'sc_project@names$cluster'}

\item{markers_subclass}{A list of subclass markers (default: NULL). If NULL the markers will used from 'sc_project@metadata$heterogeneity_markers'}
}
\value{
An updated `scRNAproject` object with new subclass names assigned. The subclass names will be stored in the `names$subclass` slot of the `sc_project`.
}
\description{
This function assigns meaningful names to subclasses and classes based on clustering and heterogeneity markers in a single-cell RNA sequencing project. It uses information about heterogeneity, class markers, and subclass markers to perform the annotation. It updates the subclass names in the `scRNAproject` object by combining cluster names with subclass names, based on the provided markers and species. This function assumes that clustering information and heterogeneity markers are already present in the `scRNAproject` metadata.
The structure of manually curated marker data, including marker class and marker subclass, is available on the GitHub page: https://github.com/jkubis96/CSSG. If set to NULL by default, the names will be based on specific markers for the clusters.
}
\details{
The function first checks if heterogeneity markers are present in the metadata. If markers are missing or invalid, it raises an error. The subclass names are then assigned by:
- Using the class markers to determine class names for the clusters.
- Using the subclass markers and heterogeneity markers to assign meaningful subclass names to cells.
- Combining the class and subclass names to form the final names for each cell.

The `aggregation_chr` and `aggregation_num` functions are used to handle different types of data (character vs numeric) during subclass aggregation.
}
\examples{
# Assuming you have a valid scRNAproject object (`sc_project`), class markers, and subclass markers
updated_project <- subclass_naming(sc_project, class_markers = NULL, subclass_markers = NULL, species = 'Homo sapiens', chunk_size = 5000)

}
