% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSSG.toolkit.R
\name{subset_project}
\alias{subset_project}
\title{Subset a Single-Cell Project Based on Specific Groupings and Select List}
\usage{
subset_project(sc_project, type = "subtypes", select_list = c())
}
\arguments{
\item{sc_project}{An object containing single-cell RNA-seq data that has been preprocessed and categorized.}

\item{type}{A string indicating the groupings to subset by. Accepted values are 'subtypes', 'subclasses', 'cluster', or 'primary'.}

\item{select_list}{A character vector of names or identifiers to be selected. Only cells belonging to these groups will be kept in the subset.}
}
\value{
A modified `sc_project` object with only the cells that match the selected groups and identifiers.
}
\description{
This function subsets a single-cell RNA-seq project object by selecting specific groups (subtypes, subclasses, clusters, or primary) based on the provided `select_list`.
Only the cells that belong to the specified `select_list` are retained in the matrices and names within the `sc_project`.
}
\examples{
subsetted_project <- subset_project(sc_project, type = "subtypes", select_list = c("Subtype1", "Subtype3"))
subsetted_project <- subset_project(sc_project, type = "cluster", select_list = c("Cluster4", "Cluster7"))
}
