% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSSG.toolkit.R
\name{subtypes_naming}
\alias{subtypes_naming}
\title{Subtypes Naming}
\usage{
subtypes_naming(
  sc_project,
  markers_class = NULL,
  markers_subclass = NULL,
  species = "Homo sapiens",
  chunk_size = 5000
)
}
\arguments{
\item{sc_project}{An object containing scRNA-seq data, including expression matrices and metadata.}

\item{markers_class}{A list of class markers (default: NULL). If NULL the cluster will obtain native names from 'sc_project@names$primary' or 'sc_project@names$cluster'}

\item{markers_subclass}{A list of subclass markers (default: NULL). If NULL the markers will used from 'sc_project@metadata$heterogeneity_markers'}

\item{species}{A character string specifying the species (default: 'Homo sapiens'), which determines the structure of the names.}

\item{chunk_size}{Numeric. Size of chunks for processing large datasets (default: 5000).}
}
\value{
The updated `sc_project` object with assigned subtypes.
}
\description{
Assigns subtype names to clusters in a scRNA-seq project (sc_project) based on markers.
}
\examples{
subtypes_naming(sc_project, markers_class, markers_subclass, species = "Homo sapiens", chunk_size = 500)

}
