% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSSG.toolkit.R
\name{outlires}
\alias{outlires}
\title{Outlier Detection and Visualization}
\usage{
outlires(input)
}
\arguments{
\item{input}{A numeric vector containing values to be analyzed.}
}
\value{
A list containing a sorted vector of detected outliers and a ggplot2 histogram visualization.
}
\description{
Identifies outliers in a numeric dataset and visualizes their distribution using a histogram.
}
\examples{
outlier_data <- outlires(c(1, 5, 10, 20, 30, 100, 150, 200))

}
